added first unit tests for calc logic
