+++ webapp/test/testsuite.qunit.html
@@ -2,11 +2,9 @@
 <html>
 
 <head>
-	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<meta charset="utf-8">
 	<title>QUnit test suite for Trasta</title>
-	<script src="../resources/sap/ui/qunit/qunit-redirect.js" 
-		data-sap-ui-testsuite="testsuite.qunit"></script>
+	<script src="../resources/sap/ui/qunit/qunit-redirect.js"></script>
+	<script src="testsuite.qunit.js" data-sap-ui-testsuite></script>
 </head>
 
 <body>
