+++ webapp/test/testsuite.qunit.js
@@ -1,15 +1,15 @@
 /* global window, parent, location */
 
 // eslint-disable-next-line
-window.suite = function() {
+window.suite = function () {
 	"use strict";
 
 	// eslint-disable-next-line
 	var oSuite = new parent.jsUnitTestSuite(),
 		sContextPath = location.pathname.substring(0, location.pathname.lastIndexOf("/") + 1);
 
-	oSuite.addTestPage(sContextPath + "integration/opaTests.qunit.html");
-	//oSuite.addTestPage(sContextPath + )
+	oSuite.addTestPage(sContextPath + "unit/unitTests.qunit.html");
+	//oSuite.addTestPage(sContextPath + "integration/opaTests.qunit.html");
 
 	return oSuite;
 };
