+++ webapp/test/unit/AllTests.js
@@ -1,0 +1,3 @@
+sap.ui.define([
+	"./libs/wastecalc"
+]);
