+++ webapp/test/unit/unitTests.qunit.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<meta charset="utf-8">
+	<title>Unit tests for Todo App</title>
+
+	<script id="sap-ui-bootstrap"
+		src="https://openui5.hana.ondemand.com/resources/sap-ui-core.js"
+		data-sap-ui-theme="sap_fiori_3_dark"
+		data-sap-ui-resourceroots='{
+			"com.fidschenberger.wasteStatsApp": "../../"
+		}'
+		data-sap-ui-language="EN"
+		data-sap-ui-async="true">
+	</script>
+
+	<link rel="stylesheet" type="text/css" href="../../resources/sap/ui/thirdparty/qunit-2.css">
+
+	<script src="../../resources/sap/ui/thirdparty/qunit-2.js"></script>
+	<script src="../../resources/sap/ui/qunit/qunit-junit.js"></script>
+	<script src="../../resources/sap/ui/qunit/qunit-coverage.js"></script>
+	<script src="../../resources/sap/ui/thirdparty/sinon.js"></script>
+	<script src="../../resources/sap/ui/thirdparty/sinon-qunit.js"></script>
+
+	<script src="unitTests.qunit.js"></script>
+</head>
+<body>
+	<div id="qunit"></div>
+	<div id="qunit-fixture"></div>
+</body>
+</html>
